==> scripts/3_bertopic.ipynb <==
#semi-guided


import pandas as pd
import torch
from bertopic import BERTopic

torch.set_num_threads(16)

grouped = pd.read_pickle('../data/large-data/grouped_df.pkl')

annotations = pd.read_csv('../data/enemy-annotations - sentences.csv')

# Take only the first value if there are multiple categories
annotations['manual_label'] = annotations['polemical category'].astype(str).str.split(',').str[0].str.strip()

# Merge with grouped on 'sentence_id'
grouped = grouped.merge(
    annotations[['sentence_id', 'manual_label']],
    on='sentence_id',
    how='left'
)

import numpy as np

grouped['manual_label'] = grouped['manual_label'].replace('nan', np.nan)
grouped['manual_label'] = grouped['manual_label'].fillna(-1).astype(int)

# Split by 'lagt_provenience'
grouped_christian = grouped[grouped['lagt_provenience'] == 'christian']
grouped_pagan = grouped[grouped['lagt_provenience'] == 'pagan']

christian_0_300 = grouped_christian[
    (grouped_christian['not_before'] >= 0) & (grouped_christian['not_after'] <= 300)
]
christian_300_600 = grouped_christian[
    (grouped_christian['not_before'] >= 300) & (grouped_christian['not_after'] <= 600)
]

pagan_0_300 = grouped_pagan[
    (grouped_pagan['not_before'] >= 0) & (grouped_pagan['not_after'] <= 300)
]
pagan_300_600 = grouped_pagan[
    (grouped_pagan['not_before'] >= 300) & (grouped_pagan['not_after'] <= 600)
]

docs = grouped['lamma_sentence']
y = grouped['manual_label']

category_names = [
    "violence, troublemaking",
    "moral depravity",
    "idolatry, heresy, magic",
    "falseness, hypocrisy, inflated self-esteem",
    "evil/dubious agents",
    "general polemical term"
]

def run_bertopic_on_subcorpus(subcorpus):
    from bertopic import BERTopic
    from transformers.pipelines import pipeline

    docs = subcorpus['lamma_sentence']
    y = subcorpus['manual_label']

    category_names = [
        "violence, troublemaking",
        "moral depravity",
        "idolatry, heresy, magic",
        "falseness, hypocrisy, inflated self-esteem",
        "evil/dubious agents",
        "general polemical term"
    ]

    embedding_model = pipeline("feature-extraction", model="pranaydeeps/Ancient-Greek-BERT")
    topic_model = BERTopic(embedding_model=embedding_model,
                            n_gram_range = (1, 1),
                            verbose=True,
                            low_memory=True
    )

    topic_model.fit(docs, y=y)
    topic_model.save(f"../data/large-data/{subcorpus}_bertopic_semi_model", serialization="safetensors", save_ctfidf=True, save_embedding_model=embedding_model)
    labeled_text_df = topic_model.get_document_info(docs)
    labeled_text_df.to_pickle(f"../data/large-data/{subcorpus}_bertopic_semi_labeled.pkl")


run_bertopic_on_subcorpus(christian_0_300)
run_bertopic_on_subcorpus(christian_300_600)
run_bertopic_on_subcorpus(pagan_0_300)
run_bertopic_on_subcorpus(pagan_300_600)

import pandas as pd

semi = pd.read_pickle('../data/large-data/christian_0_300_bertopic_semi_labeled.pkl')

import pandas as pd

semi = pd.read_pickle('../data/large-data/christian_0_300_bertopic_semi_labeled.pkl')
filtered = semi[semi['Topic'].isin(range(1, 8))]
print(len(filtered), len(semi), len(filtered)/ len(semi))

semi = pd.read_pickle('../data/large-data/christian_300_600_bertopic_semi_labeled.pkl')
filtered = semi[semi['Topic'].isin(range(1, 8))]
print(len(filtered), len(semi), len(filtered)/ len(semi))

semi = pd.read_pickle('../data/large-data/pagan_0_300_bertopic_semi_labeled.pkl')
filtered = semi[semi['Topic'].isin(range(1, 8))]
print(len(filtered), len(semi), len(filtered)/ len(semi))

semi = pd.read_pickle('../data/large-data/pagan_300_600_bertopic_semi_labeled.pkl')
filtered = semi[semi['Topic'].isin(range(1, 8))]
print(len(filtered), len(semi), len(filtered)/ len(semi))

#guided

associated terms

Violence, trouble-making
στασιαστής (rebel), ταραχώδης (troublesome), φονεύς (murderer)
Moral depravity
πόρνος (fornicator), πλεονέκτης (greedy person), μέθυσος (drunkard)
Idolatry, heresy, magic
αἱρετικός (heretic), μάγος (sorcerer, magician)
Falseness, hypocrisy
ψεύστης (liar), ὑποκριτής (hypocrite)
Inflated self-esteem
ἀλαζών (boaster), φιλοδοξία (ambition, love of glory)
Evil/dubious agents
διάβολος (devil, slanderer), κακοποιός (evildoer, criminal), πονηρός (evil), ἀντίχριστος (Antichrist), Σατανᾶς (Satan)



#zero-shot

import pandas as pd
import torch
from bertopic import BERTopic
from bertopic.representation import KeyBERTInspired
from transformers.pipelines import pipeline

grouped = pd.read_pickle('../data/large-data/grouped_df.pkl')

category_names = [
    "violence, troublemaking",
    "moral depravity",
    "idolatry, heresy, magic",
    "falseness, hypocrisy, inflated self-esteem",
    "evil/dubious agents",
    "general polemical term"
]

torch.set_num_threads(16)

embedding_model = pipeline(
    "feature-extraction",
    model="pranaydeeps/Ancient-Greek-BERT",
    #device=-1  # Use CPU
)

from umap import UMAP
umap_model = UMAP(n_neighbors=8,
                  metric='cosine',
                  n_components=2,
                  random_state=42,
                    )


#Zero-shot mode
topic_model = BERTopic(verbose=True, 
                        embedding_model=embedding_model,
                        min_topic_size=15,
                        zeroshot_topic_list=category_names,
                        zeroshot_min_similarity=0,
                        representation_model=KeyBERTInspired(),
                        umap_model=umap_model,
                        )


topics, _  = topic_model.fit_transform(grouped['lamma_sentence'])
topic_model.save("../data/large-data/topic_model")

labeled_text_df = topic_model.get_document_info(grouped['lamma_sentence'])
labeled_text_df.to_pickle("../data/large-data/labeled_text_df.pkl")

#load trained model
topic_model = BERTopic.load("../data/large-data/topic_model")
labeled_text_df.read_pickle("../data/large-data/labeled_text_df.pkl")

#print(f'{topic_model.get_topic_info()['Count'][0]/len(df)}% are unlabeled outliers.')
topic_model.get_topic_info()
